// types/lens.types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BlockchainData: any;
  BroadcastId: any;
  ChainId: any;
  CollectModuleData: any;
  ContractAddress: any;
  CreateHandle: any;
  Cursor: any;
  DateTime: any;
  Ens: any;
  EthereumAddress: any;
  FollowModuleData: any;
  Handle: any;
  HandleClaimIdScalar: any;
  InternalPublicationId: any;
  Jwt: any;
  LimitScalar: any;
  Locale: any;
  Markdown: any;
  MimeType: any;
  NftOwnershipId: any;
  Nonce: any;
  NotificationId: any;
  ProfileId: any;
  PublicationId: any;
  PublicationTag: any;
  PublicationUrl: any;
  ReferenceModuleData: any;
  Search: any;
  Signature: any;
  Sources: any;
  TimestampScalar: any;
  TxHash: any;
  TxId: any;
  UnixTimestamp: any;
  Url: any;
  Void: any;
};
// ...
export type CollectModule =
  | FeeCollectModuleSettings
  | FreeCollectModuleSettings
  | LimitedFeeCollectModuleSettings
  | LimitedTimedFeeCollectModuleSettings
  | RevertCollectModuleSettings
  | TimedFeeCollectModuleSettings;
// ...
export type Comment = {
  __typename?: "Comment";
  /** ID of the source */
  appId?: Maybe<Scalars["Sources"]>;
  /** The collect module */
  collectModule: CollectModule;
  /** The contract address for the collect nft.. if its null it means nobody collected yet as it lazy deployed */
  collectNftAddress?: Maybe<Scalars["ContractAddress"]>;
  /** Who collected it, this is used for timeline results and like this for better caching for the client */
  collectedBy?: Maybe<Wallet>;
  /** Which comment this points to if its null the pointer too deep so do another query to find it out */
  commentOn?: Maybe<Publication>;
  /** The date the post was created on */
  createdAt: Scalars["DateTime"];
  /** This will bring back the first comment of a comment and only be defined if using `publication` query and `commentOf` */
  firstComment?: Maybe<Comment>;
  hasCollectedByMe: Scalars["Boolean"];
  /** If the publication has been hidden if it has then the content and media is not available */
  hidden: Scalars["Boolean"];
  /** The internal publication id */
  id: Scalars["InternalPublicationId"];
  /** The top level post/mirror this comment lives on */
  mainPost: MainPostReference;
  /** The metadata for the post */
  metadata: MetadataOutput;
  mirrors: Array<Scalars["InternalPublicationId"]>;
  /** The on chain content uri could be `ipfs://` or `https` */
  onChainContentURI: Scalars["String"];
  /** The profile ref */
  profile: Profile;
  reaction?: Maybe<ReactionTypes>;
  /** The reference module */
  referenceModule?: Maybe<ReferenceModule>;
  /** The publication stats */
  stats: PublicationStats;
};
// ...
export type MainPostReference = Mirror | Post;
// ...
export type MetadataOutput = {
  __typename?: "MetadataOutput";
  /** The attributes */
  attributes: Array<MetadataAttributeOutput>;
  /** This is the metadata content for the publication, should be markdown */
  content?: Maybe<Scalars["Markdown"]>;
  /** The content warning for the publication */
  contentWarning?: Maybe<PublicationContentWarning>;
  /** The image cover for video/music publications */
  cover?: Maybe<MediaSet>;
  /** This is the metadata description */
  description?: Maybe<Scalars["Markdown"]>;
  /** This is the image attached to the metadata and the property used to show the NFT! */
  image?: Maybe<Scalars["Url"]>;
  /** The locale of the publication,  */
  locale?: Maybe<Scalars["Locale"]>;
  /** The main focus of the publication */
  mainContentFocus: PublicationMainFocus;
  /** The images/audios/videos for the publication */
  media: Array<MediaSet>;
  /** The metadata name */
  name?: Maybe<Scalars["String"]>;
  /** The tags for the publication */
  tags: Array<Scalars["String"]>;
};

/** The social mirror */
export type Mirror = {
  __typename?: "Mirror";
  /** ID of the source */
  appId?: Maybe<Scalars["Sources"]>;
  /** The collect module */
  collectModule: CollectModule;
  /** The contract address for the collect nft.. if its null it means nobody collected yet as it lazy deployed */
  collectNftAddress?: Maybe<Scalars["ContractAddress"]>;
  /** The date the post was created on */
  createdAt: Scalars["DateTime"];
  hasCollectedByMe: Scalars["Boolean"];
  /** If the publication has been hidden if it has then the content and media is not available */
  hidden: Scalars["Boolean"];
  /** The internal publication id */
  id: Scalars["InternalPublicationId"];
  /** The metadata for the post */
  metadata: MetadataOutput;
  /** The mirror publication */
  mirrorOf: MirrorablePublication;
  /** The on chain content uri could be `ipfs://` or `https` */
  onChainContentURI: Scalars["String"];
  /** The profile ref */
  profile: Profile;
  reaction?: Maybe<ReactionTypes>;
  /** The reference module */
  referenceModule?: Maybe<ReferenceModule>;
  /** The publication stats */
  stats: PublicationStats;
};

/** The social mirror */
export type MirrorReactionArgs = {
  request?: InputMaybe<ReactionFieldResolverRequest>;
};

export type MirrorablePublication = Comment | Post;
// ...
export type Post = {
  __typename?: "Post";
  /** ID of the source */
  appId?: Maybe<Scalars["Sources"]>;
  /** The collect module */
  collectModule: CollectModule;
  /** The contract address for the collect nft.. if its null it means nobody collected yet as it lazy deployed */
  collectNftAddress?: Maybe<Scalars["ContractAddress"]>;
  /** Who collected it, this is used for timeline results and like this for better caching for the client */
  collectedBy?: Maybe<Wallet>;
  /** The date the post was created on */
  createdAt: Scalars["DateTime"];
  hasCollectedByMe: Scalars["Boolean"];
  /** If the publication has been hidden if it has then the content and media is not available */
  hidden: Scalars["Boolean"];
  /** The internal publication id */
  id: Scalars["InternalPublicationId"];
  /** The metadata for the post */
  metadata: MetadataOutput;
  mirrors: Array<Scalars["InternalPublicationId"]>;
  /** The on chain content uri could be `ipfs://` or `https` */
  onChainContentURI: Scalars["String"];
  /** The profile ref */
  profile: Profile;
  reaction?: Maybe<ReactionTypes>;
  /** The reference module */
  referenceModule?: Maybe<ReferenceModule>;
  /** The publication stats */
  stats: PublicationStats;
};

/** The social post */
export type PostMirrorsArgs = {
  by?: InputMaybe<Scalars["ProfileId"]>;
};

/** The social post */
export type PostReactionArgs = {
  request?: InputMaybe<ReactionFieldResolverRequest>;
};

/** The Profile */
export type Profile = {
  __typename?: "Profile";
  /** Optionals param to add extra attributes on the metadata */
  attributes?: Maybe<Array<Attribute>>;
  /** Bio of the profile */
  bio?: Maybe<Scalars["String"]>;
  /** The cover picture for the profile */
  coverPicture?: Maybe<ProfileMedia>;
  /** The dispatcher */
  dispatcher?: Maybe<Dispatcher>;
  /** The follow module */
  followModule?: Maybe<FollowModule>;
  /** Follow nft address */
  followNftAddress?: Maybe<Scalars["ContractAddress"]>;
  /** The profile handle */
  handle?: Scalars["Handle"];
  /** The profile id */
  id?: Scalars["ProfileId"];
  /** Is the profile default */
  isDefault?: Scalars["Boolean"];
  isFollowedByMe?: Scalars["Boolean"];
  isFollowing?: Scalars["Boolean"];
  /** Metadata url */
  metadata?: Maybe<Scalars["Url"]>;
  /** Name of the profile */
  name?: Maybe<Scalars["String"]>;
  /** The on chain identity */
  onChainIdentity?: OnChainIdentity;
  /** Who owns the profile */
  ownedBy?: Scalars["EthereumAddress"];
  /** The picture for the profile */
  picture?: Maybe<ProfileMedia>;
  /** Profile stats */
  stats?: ProfileStats;
};
// ...
export type Publication = Comment | Mirror | Post;
// ...
export type PublicationStats = {
  __typename?: "PublicationStats";
  commentsTotal: Scalars["Int"];
  /** The publication id */
  id: Scalars["InternalPublicationId"];
  /** The total amount of collects */
  totalAmountOfCollects: Scalars["Int"];
  /** The total amount of comments */
  totalAmountOfComments: Scalars["Int"];
  /** The total amount of mirrors */
  totalAmountOfMirrors: Scalars["Int"];
  /** The total amount of upvotes */
  totalDownvotes: Scalars["Int"];
  /** The total amount of downvotes */
  totalUpvotes: Scalars["Int"];
};
// ...
export enum ReactionTypes {
  Downvote = "DOWNVOTE",
  Upvote = "UPVOTE",
}

export type ReferenceModule = FollowOnlyReferenceModuleSettings;
// ...
export type Wallet = {
  __typename?: "Wallet";
  address: Scalars["EthereumAddress"];
  /** The default profile for the wallet for now it is just their first profile, this will be the default profile they picked soon enough */
  defaultProfile?: Maybe<Profile>;
};
// ...
export enum CommentOrderingTypes {
  Desc = "DESC",
  Ranking = "RANKING",
}

/** The comment ranking filter types */
export enum CommentRankingFilter {
  NoneRelevant = "NONE_RELEVANT",
  Relevant = "RELEVANT",
}